+++ metadata.rb
@@ -9,7 +9,7 @@ source_url       'https://github.com/sous-chefs/nagios'
 chef_version     '>= 14'
 
 depends 'apache2', '>= 8.0'
-depends 'nginx', '>= 10.2'
+depends 'nginx', '>= 11.2'
 depends 'nrpe'
 depends 'php', '>= 4.0.0'
 depends 'yum-epel'